move invitation save to root app component
