+++ packages/amplication-client/src/App.tsx
@@ -1,4 +1,5 @@
 import React, { useCallback, useState, useEffect } from "react";
+import { useLocation } from "react-router-dom";
 import * as reactHotkeys from "react-hotkeys";
 import ThemeProvider from "./Layout/ThemeProvider";
 import { track, dispatch, init as initAnalytics } from "./util/analytics";
@@ -8,6 +9,10 @@ import { routesGenerator } from "./routes/routesUtil";
 import useAuthenticated from "./authentication/use-authenticated";
 import useCurrentWorkspace from "./Workspaces/hooks/useCurrentWorkspace";
 import { Loader } from "@amplication/design-system";
+import useLocalStorage from "react-use-localstorage";
+import queryString from "query-string";
+
+export const LOCAL_STORAGE_KEY_INVITATION_TOKEN = "invitationToken";
 
 const GeneratedRoutes = routesGenerator(Routes);
 const context = {
@@ -27,6 +32,8 @@ export const enhance = track<keyof typeof context>(
 
 function App() {
   const authenticated = useAuthenticated();
+  const location = useLocation();
+
   const { currentWorkspaceLoading } = useCurrentWorkspace(authenticated);
   const [keepLoadingAnimation, setKeepLoadingAnimation] =
     useState<boolean>(true);
@@ -39,6 +46,21 @@ function App() {
   const handleTimeout = useCallback(() => {
     setKeepLoadingAnimation(false);
   }, []);
+
+  const [, setInvitationToken] = useLocalStorage(
+    LOCAL_STORAGE_KEY_INVITATION_TOKEN,
+    undefined
+  );
+
+  useEffect(() => {
+    const params = queryString.parse(location.search);
+    if (params.invitation) {
+      //save the invitation token in local storage to be validated by
+      //<CompleteInvitation/> after signup or sign in
+      //we user local storage since github-passport does not support dynamic callback
+      setInvitationToken(params.invitation as string);
+    }
+  }, [setInvitationToken, location.search]);
 
   //The default behavior across all <HotKeys> components
   reactHotkeys.configure({
